refactor sds_calculation to uk_who.py. create global functions.
