+++ rcpchgrowth/rcpchgrowth/__init__.py
@@ -1,5 +1,6 @@
 from .date_calculations import decimal_age, chronological_decimal_age, corrected_decimal_age, chronological_calendar_age, estimated_date_delivery, corrected_gestational_age
-from .sds_calculations import uk_who_sds_calculation, centile, percentage_median_bmi
+from .uk_who import uk_who_sds_calculation, percentage_median_bmi
+from .global_functions import centile
 from .centile_bands import centile_band_for_centile
 from .bmi_functions import bmi_from_height_weight, weight_for_bmi_height
 from .growth_interpretations import comment_prematurity_correction
